requirements.txt
streamlit
easyocr
pillow
pandas
PyPDF2
openpyxl
PyMuPDF
